requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
psycopg2-binary==2.9.9
